refactor(settings): split AgentTab and fold on-device state into UiState
Extract OnDevice/OpenAI provider cards to keep AgentTab under the soft LOC
guideline, and move local model download/ready catalog fields into
SettingsUiState.Ready for a single Settings state surface.

Closes #463
